feat: reorganize project file structure for better clarity
